backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pinecone
sentence-transformers
langchain
langchain-openai
langchain-community
openai
pypdf
python-docx
python-pptx
openpyxl
python-dotenv
python-multipart
email-validator
requests
